+++ Source/FinancialSummary.Presentation/Api/V1/Requests/CreateDepositRequestBody.cs
@@ -3,8 +3,8 @@ namespace FinancialSummary.Presentation.Api.V1.Requests;
 public sealed record CreateDepositRequestBody(
 	Guid? OperationId,
 	string Name,
-	double Cash,
-	double InterestRate,
+	decimal Cash,
+	decimal InterestRate,
 	int CapitalizationPerYear,
 	DateTime StartDate,
 	DateTime FinishDate);
